disabled Agenda-optimization RETRACT+ASSERT=null
The optimization worked by comparing tokens by value. But when a token
(A) is retracted inside the network, and an equivalent (B) is asserted
instead, even if the resulting knowledge is the same the reasoner need
to know this to correctly track evidences (including token, compared by
ptr address!) So for now just disable this.

Should fix a bug where an explanation for a WME was missing evidences.