added abstract BetaBetaNode
BetaBetaNodes are nodes where both parents are beta memories.
My intention is to use it for a more generic noValue node that computes
the difference between two beta memories.
The design is a bit hacky, as it inherits BetaNode but ignores the
parentAlpha_ and parentBeta_ members -- it should be a very own type of
node I think. But since I doubt that it will be used for anything but
the noValue node, I guess that's okay for now.