unquote quoted strings during parsing of rules
this currently affects only strings in builtin parameters, but not in
triples, as they are explicitely specified as stringliterals, whereas
the builtin parameters can only be quoted strings. That means that
string constants when supplied to a triple still contain the quotation
marks -- which is handy, as they need to be quoted anyway.