introduced concept of AccessorConversions
AccessorConversions are a way to provide support for multiple datatypes
but convert them all to one type. This complements the accessors capabilities
to be interpreted in different ways: The accessors may not know all the
different types their data could possibly be converted too, but nodes
like e.g. InferTriple could support holding different types in a single
format - in this case rdf.

AccessorConversions is only a helper template that provides a pure
virtual interface of conversion methods that need to be implemented but
handles selecting the preferred interpretation of the given accessor to
start the conversion from, etc.