explicitely state scope of test function in is_std_to_string_valid trait
this is a workaround for a compiler bug that I confirmed to occur on g++
4.8, 4.9, 5.0 and 7.5, and is fixed in g++ 8. The bug wrongfully selects
template methods from different namespaces and classes during overload
resolution in static initialization, and thus compromises ALL external
libraries that define template classes with methods that have the same
name. In this case the rete::util::is_std_to_string_valid has a
test-method that conflicts with a test-method in a type trait in the
serialization library "cereal" used in sempr.